--- project/src/templates/js/controller.js
@@ -1,8 +1,0 @@
-export default class Controller {
-  constructor(view, model) {
-    this._view = view;
-    this._model = model;
-
-  }
-
-}
