--- project/src/templates/js/dima.js
@@ -1,9 +1,0 @@
-export default class TestGitHub {
-	constructor(text) {
-		this.text = text;
-	}
-
-	consoleDimaText(){
-		console.log(this.text)
-	}
-}
