--- project/src/templates/js/model.js
@@ -1,7 +1,0 @@
-import axios from 'axios';
-export default class Model {
-  constructor(items = []) {
-    this._items = items;
-  }
-
-}
