--- project/src/templates/sass/styles.scss
@@ -1,4 +1,0 @@
-body {
-  background-color: rgb(232, 240, 239);
-}
-
